Remove some test files; parallelize library tests.
