+++ kifi-backend/modules/shoebox/angular/test/e2e/library-private.spec.js
@@ -1,0 +1,28 @@
+describe('kifi angular library test suite', function () {
+  'use strict';
+
+  var util = require('./util/util');
+
+  describe('Kifi private library page for logged in users', function () {
+    beforeEach(function () {
+      browser.get('/integrator-aa/secret');
+      util.loginUserWithCookies();
+    });
+
+    afterEach(function () {
+      util.checkJSErrors();
+    });
+
+    it('should have the correct title', function () {
+      util.checkForCorrectTitle('My Private Library by Integrator AA • Kifi');
+    });
+
+    it('should display the library card header', function () {
+      util.checkForLibraryCardHeader();
+    });
+
+    it('should display the correct library name', function () {
+      util.checkForCorrectLibraryName('My Private Library');
+    });
+  });
+});
